move ToolAdapter and ToolDispatcher to Legion::LLM::Tools namespace
- canonical location now lib/legion/llm/tools/{adapter,dispatcher}.rb
- pipeline/ files are backwards-compat shims (Pipeline::ToolAdapter = Tools::Adapter)
- McpToolAdapter alias preserved
- all 1680 specs passing unchanged
